requirements.txt
streamlit
thefuzz
python-Levenshtein
pytesseract
Pillow
fpdf2
chromadb
sentence-transformers
transformers
torch
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
thefuzz>=0.20.0
python-Levenshtein>=0.21.0
opencv-python-headless
typer
rich
fastapi
uvicorn
pydantic

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
